pyproject.toml
[tool.poetry]
name = "customer-support-assistant"
version = "0.1.0"
description = "A FastAPI-based customer support assistant with AI integration"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
pydantic = "^2.4.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aiohttp = "^3.8.6"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
alembic = "^1.12.1"
uuid = "^1.30"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 